integration-tests: refactor
