+++ integration-tests/cli/test/fixtures/output.json
@@ -1,0 +1,5 @@
+{
+  "data": {
+    "count": 12
+  }
+}
