+++ integration-tests/cli/test/fixtures/simple.js
